chore(typescript): migrate `victory-accessible-group`
